Add new nginx lb for public services
For services we want exposed to the Internet and that also want to use
the toothyshouse.com domain.

Within the network, we can access the services using domains that
resolve to 172.16.20.29:80. Outside of the local network, we'll use domains
that resolve to our public IP and use inbound NAT configuration to set
up port forwarding to 172.16.20.29:80. The external port may be
different.